+++ src/youandme/commands.py
@@ -1,1 +1,2 @@
 garbage_character = b'\x00'
+WELCOME_MESSAGE = "Connection established. Dance like no one is watching :)\n"