+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 This project uses Semantic Versioning
 
+
+## 1.1.1
+
+Fixed setup.py project URL
+
+
+## 1.1.0
+
+Added detection of connection loss
+Added --alphanum-only argument
+
 ## 1.0.0
 
 Made yam a proper script (install and you can run from path with $ yam.py <args>)